+++ tests/PH.DbAppSettings.Tests/Helpers/DbContextHelper.cs
@@ -1,35 +1,37 @@
 using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
-using PH.DbAppSettings.Data;
 
 namespace PH.DbAppSettings.Tests.Helpers;
 
 public static class DbContextHelper
 {
-    public static AppSettingsDbContext CreateInMemoryContext(SqliteConnection? connection = null)
+    public static TestAppSettingsDbContext CreateInMemoryContext(SqliteConnection? connection = null)
     {
         var conn = connection ?? new SqliteConnection("Data Source=:memory:");
-        conn.Open();
+        if (conn.State != System.Data.ConnectionState.Open)
+        {
+            conn.Open();
+        }
 
-        var options = new DbContextOptionsBuilder<AppSettingsDbContext>()
+        var options = new DbContextOptionsBuilder<TestAppSettingsDbContext>()
             .UseSqlite(conn)
             .Options;
 
-        var context = new AppSettingsDbContext(options);
+        var context = new TestAppSettingsDbContext(options);
         context.Database.EnsureCreated();
         return context;
     }
 
-    public static (AppSettingsDbContext context, SqliteConnection connection) CreateSharedInMemoryContext()
+    public static (TestAppSettingsDbContext context, SqliteConnection connection) CreateSharedInMemoryContext()
     {
         var connection = new SqliteConnection("Data Source=:memory:");
         connection.Open();
 
-        var options = new DbContextOptionsBuilder<AppSettingsDbContext>()
+        var options = new DbContextOptionsBuilder<TestAppSettingsDbContext>()
             .UseSqlite(connection)
             .Options;
 
-        var context = new AppSettingsDbContext(options);
+        var context = new TestAppSettingsDbContext(options);
         context.Database.EnsureCreated();
         return (context, connection);
     }